% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directHaus.R
\name{directHaus}
\alias{directHaus}
\title{Calculate the directional extended Hausdorff Distance}
\usage{
directHaus(A, B, f1, f2, overlap = F)
}
\arguments{
\item{A, B}{region calculate the extended Hausdorff distance- SpatialPolygons or SpatialPolygonsDataFrame}

\item{f1}{the percentage of area in B you want captured as a decimal (eg 10\% = .1)}

\item{f2}{the percentage of area in A you want captured as a decimal (eg 10\% = .1)}

\item{overlap}{return the portion of A used to calculate the directional Hausdorff distance? Defaults to FALSE}
}
\value{
The directional extended hausdorff distance A to B
}
\description{
This function takes a SpatialPolygons object
}
\author{
Julia Schedler
}
