% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extHaus.R
\name{extHaus}
\alias{extHaus}
\title{Extended Hausdorff Distance}
\usage{
extHaus(A, B, f1, f2 = f1, pres = 3)
}
\arguments{
\item{A, B}{region calculate the extended Hausdorff distance- SpatialPolygons or SpatialPolygonsDataFrame}

\item{f1}{the percentage of area in B you want captured as a decimal (eg 10\% = .1)}

\item{f2}{the percentage of area in A you want captured as a decimal (eg 10\% = .1)}

\item{pres}{the number of decimal places/precision}
}
\value{
haus.dist: the extended hausdorff distance (max of directional from A to B and B to A)
}
\description{
Calculates the extended Hausdorff distance from A to B for given quantiles and precision
}
\author{
Julia Schedler
}
