% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairHaus.R
\name{pairHaus}
\alias{pairHaus}
\title{Pairwise Hausdorff distances}
\usage{
pairHaus(shp, vals = class(shp) == "SpatialPolygonsDataFrame", f1,
  f2 = f1, col)
}
\arguments{
\item{shp}{a SpatialPolygonsDataFrame object.}
}
\value{

}
\description{
This function takes a SpatialPolygonsDataFrame object
}
