% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgmHaus.R
\name{vgm_haus}
\alias{vgm_haus}
\title{Create a variogram based on Hausdorff distances}
\usage{
vgm_haus(shp, col, f1, f2 = f1)
}
\arguments{
\item{shp}{a SpatialPolygonsDataFrame object.}
}
\value{
an nxn matrix of requested distances or inverse.
}
\description{
This function takes a SpatialPolygonsDataFrame object
}
