% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgmHaus2.R
\name{vgmHaus2}
\alias{vgmHaus2}
\title{Creating a Variogram from Hausdorff Distances}
\usage{
vgmHaus2(shp, hausMat = NA, cut = "max", nbins = 15,
  vg.model = "Exp", outcome, ...)
}
\arguments{
\item{shp}{a SpatialPolygons object.}

\item{hausMat}{a weight matrix outputted by \code{\link{hausMat}}}

\item{cut}{"bbox" or "max" method of specifying distance cutoff}

\item{...}{arguments to be passed to \code{\link{hausMat}}}

\item{bbox}{if no shp is specified, the bbox of your shapefile}

\item{z}{variable for which to compute variogram}

\item{model}{parametric form of the variogram model to be estimated; defaults to "Exp".}
}
\value{
an empirical variogram with estimated parameters.
}
\description{
This function takes a SpatialPolygons object
}
